+++ src/main/java/net/mohron/skyclaims/command/CommandBase.java
@@ -16,20 +16,13 @@
  * along with SkyClaims.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.mohron.skyclaims.claim;
+package net.mohron.skyclaims.command;
 
-import com.flowpowered.math.vector.Vector3i;
-import me.ryanhamshire.griefprevention.api.claim.Claim;
-import me.ryanhamshire.griefprevention.api.claim.ClaimResult;
-import me.ryanhamshire.griefprevention.api.claim.ClaimType;
-import org.spongepowered.api.entity.living.player.Player;
-import org.spongepowered.api.world.World;
+import net.mohron.skyclaims.SkyClaims;
+import org.spongepowered.api.command.spec.CommandExecutor;
 
-import java.util.Optional;
-import java.util.UUID;
+public abstract class CommandBase implements CommandExecutor {
 
-public interface IClaimSystem {
-	public ClaimResult createClaim(World world, Vector3i a, Vector3i b, UUID claimId, Claim parent, ClaimType claimType, boolean cuboid, Player player);
+	protected static final SkyClaims PLUGIN = SkyClaims.getInstance();
 
-	public Optional<Claim> getClaim(UUID claimId);
 }