+++ src/main/java/net/mohron/skyclaims/command/admin/CommandCreateSchematic.java
@@ -18,8 +18,10 @@
 
 package net.mohron.skyclaims.command.admin;
 
+import static org.spongepowered.api.command.args.GenericArguments.string;
+
 import com.flowpowered.math.vector.Vector3i;
-import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.command.argument.SchematicArgument;
 import net.mohron.skyclaims.listener.SchematicHandler;
 import net.mohron.skyclaims.permissions.Permissions;
@@ -27,7 +29,6 @@ import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
-import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.data.DataContainer;
 import org.spongepowered.api.data.persistence.DataFormats;
@@ -43,10 +44,8 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.util.zip.GZIPOutputStream;
 
-import static org.spongepowered.api.command.args.GenericArguments.string;
+public class CommandCreateSchematic extends CommandBase {
 
-public class CommandCreateSchematic implements CommandExecutor {
-	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
 	public static final String HELP_TEXT = "used to save the selected area as an island schematic";
 	private static final Text NAME = Text.of("name");