+++ src/main/java/net/mohron/skyclaims/command/user/CommandSetSpawn.java
@@ -18,21 +18,20 @@
 
 package net.mohron.skyclaims.command.user;
 
-import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.command.CommandBase;
 import net.mohron.skyclaims.permissions.Permissions;
 import net.mohron.skyclaims.world.Island;
 import org.spongepowered.api.command.CommandException;
 import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.CommandSource;
 import org.spongepowered.api.command.args.CommandContext;
-import org.spongepowered.api.command.spec.CommandExecutor;
 import org.spongepowered.api.command.spec.CommandSpec;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.text.Text;
 import org.spongepowered.api.text.format.TextColors;
 
-public class CommandSetSpawn implements CommandExecutor {
-	private static final SkyClaims PLUGIN = SkyClaims.getInstance();
+public class CommandSetSpawn extends CommandBase {
+
 	public static final String HELP_TEXT = "set your spawn location for your island.";
 
 	public static CommandSpec commandSpec = CommandSpec.builder()
@@ -52,15 +51,17 @@ public class CommandSetSpawn implements CommandExecutor {
 	}
 
 	public CommandResult execute(CommandSource src, CommandContext args) throws CommandException {
-		if (!(src instanceof Player))
+		if (!(src instanceof Player)) {
 			throw new CommandException(Text.of("You must be a player to use this command!"));
+		}
 
 		Player player = (Player) src;
 		Island island = Island.get(player.getLocation())
 			.orElseThrow(() -> new CommandException(Text.of("You must be on an island to use this command!")));
 
-		if (!island.getOwnerUniqueId().equals(player.getUniqueId()) && !player.hasPermission(Permissions.COMMAND_SET_SPAWN_OTHERS))
+		if (!island.getOwnerUniqueId().equals(player.getUniqueId()) && !player.hasPermission(Permissions.COMMAND_SET_SPAWN_OTHERS)) {
 			throw new CommandException(Text.of("Only the island owner may use this command!"));
+		}
 
 		island.setSpawn(player.getTransform());
 		player.sendMessage(Text.of("Your island spawn has been set to ", TextColors.GRAY, "(",