+++ src/main/java/net/mohron/skyclaims/config/type/IntegrationConfig.java
@@ -18,39 +18,17 @@
 
 package net.mohron.skyclaims.config.type;
 
-import net.mohron.skyclaims.SkyClaims;
+import net.mohron.skyclaims.config.type.integration.NucleusConfig;
 import ninja.leaping.configurate.objectmapping.Setting;
 import ninja.leaping.configurate.objectmapping.serialize.ConfigSerializable;
 
-import java.io.File;
-
 @ConfigSerializable
-public class StorageConfig {
+public class IntegrationConfig {
 
-	@Setting(value = "Location", comment = "The location to store SkyClaims data. Default: ${CONFIG}/data")
-	private String location = "${CONFIG}/data";
-	@Setting(value = "Type", comment = "The type of data storage to use. Supports [SQLITE, MYSQL]")
-	private StorageType type = StorageType.SQLITE;
-	@Setting(value = "MySQL")
-	private MysqlConfig mysqlConfig = new MysqlConfig();
+	@Setting(value = "Nucleus")
+	private NucleusConfig nucleus = new NucleusConfig();
 
-	public String getLocation() {
-		return location
-			.replace("*CONFIG*", SkyClaims.getInstance().getConfigDir().toString())
-			.replace("${CONFIG}", SkyClaims.getInstance().getConfigDir().toString())
-			.replace("/", File.separator)
-			.replace("\\", File.separator);
-	}
-
-	public StorageType getType() {
-		return type;
-	}
-
-	public void setType(StorageType type) {
-		this.type = type;
-	}
-
-	public MysqlConfig getMysqlConfig() {
-		return mysqlConfig;
+	public NucleusConfig getNucleus() {
+		return nucleus;
 	}
 }